+++ scripts/flow/config/flowconfig
@@ -20,6 +20,7 @@
 [libs]
 ../../../node_modules/fbjs/flow/lib/dev.js
 ../environment.js
+../react-native-host-hooks.js
 
 [lints]
 untyped-type-import=error
